+++ library/Mockery/Matcher/Subset.php
@@ -81,12 +81,21 @@ class Subset extends MatcherAbstract
      */
     public function __toString()
     {
-        $return = '<Subset[';
-        $elements = array();
-        foreach ($this->expected as $k=>$v) {
-            $elements[] = $k . '=' . (string) $v;
+        return '<Subset' . $this->formatArray($this->expected) . ">";
+    }
+
+    /**
+     * Recursively format an array into the string representation for this matcher
+     *
+     * @param array $array
+     * @return string
+     */
+    protected function formatArray(array $array)
+    {
+        $elements = [];
+        foreach ($array as $k => $v) {
+            $elements[] = $k . '=' . (is_array($v) ? $this->formatArray($v) : (string) $v);
         }
-        $return .= implode(', ', $elements) . ']>';
-        return $return;
+        return "[" . implode(", ", $elements) . "]";
     }
 }