+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * [#1245 Deprecate hamcrest/hamcrest-php package](https://github.com/mockery/mockery/pull/1245)
 * [#1246 Add BUG_REPORT.yml Issue template](https://github.com/mockery/mockery/pull/1246)
 * [#1250 Deprecate PHP <=8.0](https://github.com/mockery/mockery/issues/1250)
+* [#1253 Prevent array to string conversion when serialising a Subset matcher](https://github.com/mockery/mockery/issues/1253)
 
 ## 1.3.6 (2022-09-07)
 * PHP 8.2 | Fix "Use of "parent" in callables is deprecated" notice #1169